remove duplication from sqlite platform
