+++ packages/knex/src/dialects/index.ts
@@ -1,4 +1,1 @@
-export * from './SqliteTableCompiler';
-export * from './BetterSqliteKnexDialect';
-export * from './LibSqlKnexDialect';
-export * from './SqliteKnexDialect';
+export * from './sqlite';
