+++ packages/knex/src/dialects/sqlite/index.ts
@@ -1,0 +1,5 @@
+export * from './BaseSqlitePlatform';
+export * from './SqliteTableCompiler';
+export * from './BetterSqliteKnexDialect';
+export * from './LibSqlKnexDialect';
+export * from './SqliteKnexDialect';
